% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{summary.neural_network}
\alias{summary.neural_network}
\title{Summary method for the nn function}
\usage{
\method{summary}{neural_network}(object, prn = TRUE, ...)
}
\arguments{
\item{object}{Return value from \code{\link{nn}}}

\item{prn}{Print list of weights}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the nn function
}
\details{
See \url{https://radiant-rstats.github.io/docs/model/nn.html} for an example in Radiant
}
